src/requirements.txt
langgraph==0.4.3
pandas==2.2.3
python-telegram-bot==22.0
openai==1.78.1
python-dotenv==1.1.0
aiohttp==3.12.0
nest_asyncio==1.6.0
python-abc==0.2.0
GigaChat==0.1.39.post1